ControllerView created for TicTacToe
Created the controller view. Still need to make up my junit tests for
it, parts of it are going to need added once I know exactly how the
strategies are being developed but until then it serves its purpose.
Added a little bit to the Model to support stroking a line along the
winning row/col/diagonal